+++ src/Itau.AutoInvest.Infrastructure/Tables/ClientsTable.cs
@@ -10,7 +10,7 @@ public class ClientsTable
 {
     [Key]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-    public BigInteger Id { get; set; }
+    public long Id { get; set; }
 
     [Required]
     [Column("Nome")]
